--- Front/app/src/main/java/ItemDatabase.java
@@ -1,25 +1,0 @@
-//import android.content.Context;
-//import android.database.sqlite.SQLiteDatabase;
-//import android.database.sqlite.SQLiteOpenHelper;
-//
-//public class TodoItemDatabase extends SQLiteOpenHelper {
-//    public TodoItemDatabase(Context context) {
-//        super(context, DATABASE_NAME, null, DATABASE_VERSION);
-//    }
-//
-//    // This is where we need to write create table statements.
-//    // This is called when database is created.
-//    @Override
-//    public void onCreate(SQLiteDatabase db) {
-//        // SQL for creating the tables
-//    }
-//
-//    // This method is called when database is upgraded like
-//    // modifying the table structure,
-//    // adding constraints to database, etc
-//    @Override
-//    public void onUpgrade(SQLiteDatabase db, int oldVersion,
-//                          int newVersion) {
-//        // SQL for upgrading the tables
-//    }
-//}
