+++ Front/app/src/main/java/com/checkinn/front/activities/LoginActivity.java
@@ -1,9 +1,8 @@
-package com.checkinn.front;
+package com.checkinn.front.activities;
 
 import android.animation.Animator;
 import android.animation.AnimatorListenerAdapter;
 import android.annotation.TargetApi;
-import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.support.annotation.NonNull;
 import android.support.design.widget.Snackbar;
@@ -29,6 +28,8 @@ import android.widget.AutoCompleteTextView;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
+
+import com.checkinn.front.R;
 
 import java.util.ArrayList;
 import java.util.List;
