+++ Front/app/src/main/java/com/checkinn/front/activities/MainActivity.java
@@ -1,5 +1,6 @@
-package com.checkinn.front;
+package com.checkinn.front.activities;
 
+import android.arch.persistence.room.Room;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -7,6 +8,10 @@ import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
+
+import com.checkinn.front.R;
+import com.checkinn.front.database.database.AppDatabase;
+import com.checkinn.front.database.entities.Item;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -25,6 +30,18 @@ public class MainActivity extends AppCompatActivity {
 //        }
         startLogin();
 
+
+        //bad practice running on main thread but whatever
+        AppDatabase db = Room.databaseBuilder(getApplicationContext(),
+                AppDatabase.class, "database").allowMainThreadQueries().build();
+
+        //testing database access
+        Item item = new Item();
+        item.shopName = "shop";
+        item.itemName = "amazing_item";
+        item.price = 25.15;
+        db.userDao().insertItems(item);
+        Item[] items = db.userDao().loadAllItems();
 
         //login button for camera
         FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
