+++ Front/app/src/main/res/layout/activity_login.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:background="@drawable/gradient"
-    tools:context="com.checkinn.front.LoginActivity">
+    tools:context="com.checkinn.front.activities.LoginActivity">
     <!-- Login progress -->
     <ProgressBar
         android:id="@+id/login_progress"
