+++ Front/app/src/main/res/layout/content_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.checkinn.front.MainActivity"
+    tools:context="com.checkinn.front.activities.MainActivity"
     tools:showIn="@layout/activity_main">
 
     <ExpandableListView
